// src/domain/auth/schemas/signup.schema.ts
import { z } from 'zod';

export const signupSchema = z.object({
  name: z.string().min(2, 'Name must be at least 2 characters'),
  email: z.string().email('Please enter a valid email address'),
  password: z.string().min(6, 'Password must be at least 6 characters'),
  company: z.string().optional()
});

export type TSignupSchema = z.infer<typeof signupSchema>;